sonarlint扫描修复 修复Response问题

+++ wind-server/wind-web/wind-vue/src/main/java/com/sunseagear/wind/common/helper/JWTHelper.java
@@ -84,11 +84,11 @@ public class JWTHelper {
     }
 
     public static long getExpireIn() {
-        return Integer.parseInt(SysConfigHelper.getInstance().getSysConfig("tokenExpire").getValue()) * 60 * 1000;
+        return (long) (long) Integer.parseInt(SysConfigHelper.getInstance().getSysConfig("tokenExpire").getValue()) * 60 * 1000;
     }
 
     public static long getRefreshExpireIn() {
-        return Integer.parseInt(SysConfigHelper.getInstance().getSysConfig("refreshTokenExpire").getValue()) * 60 * 1000;
+        return (long) (long) Integer.parseInt(SysConfigHelper.getInstance().getSysConfig("refreshTokenExpire").getValue()) * 60 * 1000;
     }
 
 }
